# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Robot Fleet API
  description: Public API for QA engineers to practice API testing, automation,
    contract testing, performance testing, security testing and modern backend
    quality engineering.
  version: 1.0.0
  contact:
    name: Robot Fleet API
    url: https://github.com
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    def-0:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required:
        - page
        - limit
        - total
        - totalPages
      title: PaginationMeta
...
paths:
...
  /api/v1/jobs/:
    get:
      summary: List jobs
      tags:
        - Jobs
      parameters:
        - schema:
            minimum: 1
            default: 1
            type: integer
          in: query
          name: page
          required: false
        - schema:
            minimum: 1
            maximum: 100
            default: 20
            type: integer
          in: query
          name: limit
          required: false
        - schema:
            anyOf:
              - type: string
                enum:
                  - CREATED
              - type: string
                enum:
                  - ASSIGNED
              - type: string
                enum:
                  - RUNNING
              - type: string
                enum:
                  - PAUSED
              - type: string
                enum:
                  - COMPLETED
              - type: string
                enum:
                  - FAILED
              - type: string
                enum:
                  - CANCELLED
          in: query
          name: status
          required: false
        - schema:
            anyOf:
              - type: string
                enum:
                  - LOW
              - type: string
                enum:
                  - MEDIUM
              - type: string
                enum:
                  - HIGH
              - type: string
                enum:
                  - CRITICAL
          in: query
          name: priority
          required: false
        - schema:
            format: uuid
            type: string
          in: query
          name: robotId
          required: false
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          format: uuid
                          type: string
                        title:
                          type: string
                        description:
                          anyOf:
                            - type: string
                            - type: "null"
                        priority:
                          anyOf:
                            - type: string
                              enum:
                                - LOW
                            - type: string
                              enum:
                                - MEDIUM
                            - type: string
                              enum:
                                - HIGH
                            - type: string
                              enum:
                                - CRITICAL
                        status:
                          anyOf:
                            - type: string
                              enum:
                                - CREATED
                            - type: string
                              enum:
                                - ASSIGNED
                            - type: string
                              enum:
                                - RUNNING
                            - type: string
                              enum:
                                - PAUSED
                            - type: string
                              enum:
                                - COMPLETED
                            - type: string
                              enum:
                                - FAILED
                            - type: string
                              enum:
                                - CANCELLED
                        robotId:
                          anyOf:
                            - format: uuid
                              type: string
                            - type: "null"
                        createdBy:
                          format: uuid
                          type: string
                        scheduledAt:
                          anyOf:
                            - format: date-time
                              type: string
                            - type: "null"
                        startedAt:
                          anyOf:
                            - format: date-time
                              type: string
                            - type: "null"
                        completedAt:
                          anyOf:
                            - format: date-time
                              type: string
                            - type: "null"
                        createdAt:
                          format: date-time
                          type: string
                        updatedAt:
                          format: date-time
                          type: string
                      required:
                        - id
                        - title
                        - description
                        - priority
                        - status
                        - robotId
                        - createdBy
                        - scheduledAt
                        - startedAt
                        - completedAt
                        - createdAt
                        - updatedAt
                      title: Job
                  pagination:
                    $ref: "#/components/schemas/def-0"
                required:
                  - data
                  - pagination
    post:
      summary: Create a job
      tags:
        - Jobs
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  minLength: 1
                  type: string
                description:
                  type: string
                priority:
                  anyOf:
                    - type: string
                      enum:
                        - LOW
                    - type: string
                      enum:
                        - MEDIUM
                    - type: string
                      enum:
                        - HIGH
                    - type: string
                      enum:
                        - CRITICAL
                scheduledAt:
                  format: date-time
                  type: string
              required:
                - title
        required: true
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    format: uuid
                    type: string
                  title:
                    type: string
                  description:
                    anyOf:
                      - type: string
                      - type: "null"
                  priority:
                    anyOf:
                      - type: string
                        enum:
                          - LOW
                      - type: string
                        enum:
                          - MEDIUM
                      - type: string
                        enum:
                          - HIGH
                      - type: string
                        enum:
                          - CRITICAL
                  status:
                    anyOf:
                      - type: string
                        enum:
                          - CREATED
                      - type: string
                        enum:
                          - ASSIGNED
                      - type: string
                        enum:
                          - RUNNING
                      - type: string
                        enum:
                          - PAUSED
                      - type: string
                        enum:
                          - COMPLETED
                      - type: string
                        enum:
                          - FAILED
                      - type: string
                        enum:
                          - CANCELLED
                  robotId:
                    anyOf:
                      - format: uuid
                        type: string
                      - type: "null"
                  createdBy:
                    format: uuid
                    type: string
                  scheduledAt:
                    anyOf:
                      - format: date-time
                        type: string
                      - type: "null"
                  startedAt:
                    anyOf:
                      - format: date-time
                        type: string
                      - type: "null"
                  completedAt:
                    anyOf:
                      - format: date-time
                        type: string
                      - type: "null"
                  createdAt:
                    format: date-time
                    type: string
                  updatedAt:
                    format: date-time
                    type: string
                required:
                  - id
                  - title
                  - description
                  - priority
                  - status
                  - robotId
                  - createdBy
                  - scheduledAt
                  - startedAt
                  - completedAt
                  - createdAt
                  - updatedAt
  /api/v1/jobs/{id}:
    get:
      summary: Get a job by id
      tags:
        - Jobs
      parameters:
        - schema:
            format: uuid
            type: string
          in: path
          name: id
          required: true
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    format: uuid
                    type: string
                  title:
                    type: string
                  description:
                    anyOf:
                      - type: string
                      - type: "null"
                  priority:
                    anyOf:
                      - type: string
                        enum:
                          - LOW
                      - type: string
                        enum:
                          - MEDIUM
                      - type: string
                        enum:
                          - HIGH
                      - type: string
                        enum:
                          - CRITICAL
                  status:
                    anyOf:
                      - type: string
                        enum:
                          - CREATED
                      - type: string
                        enum:
                          - ASSIGNED
                      - type: string
                        enum:
                          - RUNNING
                      - type: string
                        enum:
                          - PAUSED
                      - type: string
                        enum:
                          - COMPLETED
                      - type: string
                        enum:
                          - FAILED
                      - type: string
                        enum:
                          - CANCELLED
                  robotId:
                    anyOf:
                      - format: uuid
                        type: string
                      - type: "null"
                  createdBy:
                    format: uuid
                    type: string
                  scheduledAt:
                    anyOf:
                      - format: date-time
                        type: string
                      - type: "null"
                  startedAt:
                    anyOf:
                      - format: date-time
                        type: string
                      - type: "null"
                  completedAt:
                    anyOf:
                      - format: date-time
                        type: string
                      - type: "null"
                  createdAt:
                    format: date-time
                    type: string
                  updatedAt:
                    format: date-time
                    type: string
                required:
                  - id
                  - title
                  - description
                  - priority
                  - status
                  - robotId
                  - createdBy
                  - scheduledAt
                  - startedAt
                  - completedAt
                  - createdAt
                  - updatedAt
    put:
      summary: Replace a job
      tags:
        - Jobs
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  minLength: 1
                  type: string
                description:
                  type: string
                priority:
                  anyOf:
                    - type: string
                      enum:
                        - LOW
                    - type: string
                      enum:
                        - MEDIUM
                    - type: string
                      enum:
                        - HIGH
                    - type: string
                      enum:
                        - CRITICAL
                scheduledAt:
                  format: date-time
                  type: string
              required:
                - title
                - priority
        required: true
      parameters:
        - schema:
            format: uuid
            type: string
          in: path
          name: id
          required: true
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    format: uuid
                    type: string
                  title:
                    type: string
                  description:
                    anyOf:
                      - type: string
                      - type: "null"
                  priority:
                    anyOf:
                      - type: string
                        enum:
                          - LOW
                      - type: string
                        enum:
                          - MEDIUM
                      - type: string
                        enum:
                          - HIGH
                      - type: string
                        enum:
                          - CRITICAL
                  status:
                    anyOf:
                      - type: string
                        enum:
                          - CREATED
                      - type: string
                        enum:
                          - ASSIGNED
                      - type: string
                        enum:
                          - RUNNING
                      - type: string
                        enum:
                          - PAUSED
                      - type: string
                        enum:
                          - COMPLETED
                      - type: string
                        enum:
                          - FAILED
                      - type: string
                        enum:
                          - CANCELLED
                  robotId:
                    anyOf:
                      - format: uuid
                        type: string
                      - type: "null"
                  createdBy:
                    format: uuid
                    type: string
                  scheduledAt:
                    anyOf:
                      - format: date-time
                        type: string
                      - type: "null"
                  startedAt:
                    anyOf:
                      - format: date-time
                        type: string
                      - type: "null"
                  completedAt:
                    anyOf:
                      - format: date-time
                        type: string
                      - type: "null"
                  createdAt:
                    format: date-time
                    type: string
                  updatedAt:
                    format: date-time
                    type: string
                required:
                  - id
                  - title
                  - description
                  - priority
                  - status
                  - robotId
                  - createdBy
                  - scheduledAt
                  - startedAt
                  - completedAt
                  - createdAt
                  - updatedAt
    patch:
      summary: Partially update a job
      tags:
        - Jobs
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  minLength: 1
                  type: string
                description:
                  type: string
                priority:
                  anyOf:
                    - type: string
                      enum:
                        - LOW
                    - type: string
                      enum:
                        - MEDIUM
                    - type: string
                      enum:
                        - HIGH
                    - type: string
                      enum:
                        - CRITICAL
                scheduledAt:
                  format: date-time
                  type: string
      parameters:
        - schema:
            format: uuid
            type: string
          in: path
          name: id
          required: true
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    format: uuid
                    type: string
                  title:
                    type: string
                  description:
                    anyOf:
                      - type: string
                      - type: "null"
                  priority:
                    anyOf:
                      - type: string
                        enum:
                          - LOW
                      - type: string
                        enum:
                          - MEDIUM
                      - type: string
                        enum:
                          - HIGH
                      - type: string
                        enum:
                          - CRITICAL
                  status:
                    anyOf:
                      - type: string
                        enum:
                          - CREATED
                      - type: string
                        enum:
                          - ASSIGNED
                      - type: string
                        enum:
                          - RUNNING
                      - type: string
                        enum:
                          - PAUSED
                      - type: string
                        enum:
                          - COMPLETED
                      - type: string
                        enum:
                          - FAILED
                      - type: string
                        enum:
                          - CANCELLED
                  robotId:
                    anyOf:
                      - format: uuid
                        type: string
                      - type: "null"
                  createdBy:
                    format: uuid
                    type: string
                  scheduledAt:
                    anyOf:
                      - format: date-time
                        type: string
                      - type: "null"
                  startedAt:
                    anyOf:
                      - format: date-time
                        type: string
                      - type: "null"
                  completedAt:
                    anyOf:
                      - format: date-time
                        type: string
                      - type: "null"
                  createdAt:
                    format: date-time
                    type: string
                  updatedAt:
                    format: date-time
                    type: string
                required:
                  - id
                  - title
                  - description
                  - priority
                  - status
                  - robotId
                  - createdBy
                  - scheduledAt
                  - startedAt
                  - completedAt
                  - createdAt
                  - updatedAt
    delete:
      summary: Delete a job
      tags:
        - Jobs
      parameters:
        - schema:
            format: uuid
            type: string
          in: path
          name: id
          required: true
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Default Response
  /api/v1/jobs/{id}/assign:
    post:
      summary: Assign a robot to a job (concurrency-safe)
      tags:
        - Jobs
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                robotId:
                  format: uuid
                  type: string
              required:
                - robotId
        required: true
      parameters:
        - schema:
            format: uuid
            type: string
          in: path
          name: id
          required: true
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    format: uuid
                    type: string
                  title:
                    type: string
                  description:
                    anyOf:
                      - type: string
                      - type: "null"
                  priority:
                    anyOf:
                      - type: string
                        enum:
                          - LOW
                      - type: string
                        enum:
                          - MEDIUM
                      - type: string
                        enum:
                          - HIGH
                      - type: string
                        enum:
                          - CRITICAL
                  status:
                    anyOf:
                      - type: string
                        enum:
                          - CREATED
                      - type: string
                        enum:
                          - ASSIGNED
                      - type: string
                        enum:
                          - RUNNING
                      - type: string
                        enum:
                          - PAUSED
                      - type: string
                        enum:
                          - COMPLETED
                      - type: string
                        enum:
                          - FAILED
                      - type: string
                        enum:
                          - CANCELLED
                  robotId:
                    anyOf:
                      - format: uuid
                        type: string
                      - type: "null"
                  createdBy:
                    format: uuid
                    type: string
                  scheduledAt:
                    anyOf:
                      - format: date-time
                        type: string
                      - type: "null"
                  startedAt:
                    anyOf:
                      - format: date-time
                        type: string
                      - type: "null"
                  completedAt:
                    anyOf:
                      - format: date-time
                        type: string
                      - type: "null"
                  createdAt:
                    format: date-time
                    type: string
                  updatedAt:
                    format: date-time
                    type: string
                required:
                  - id
                  - title
                  - description
                  - priority
                  - status
                  - robotId
                  - createdBy
                  - scheduledAt
                  - startedAt
                  - completedAt
                  - createdAt
                  - updatedAt
...
servers:
  - url: /
    description: Current server
...